New data dump using insert statements.  Has the long entrez gene names
psql has been acting "funny" when loading the data, and by "funny" I mean it
tends to ignore entire tables that exist in the plain text dump for no
particular reason.  Importing the data using 'import' statements seems to
circumvent these problems, but the file is huge and slow.  This is a stopgap
until we figure out why this is occuring.